requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dlib @ file:///C:/Users/Pc/Desktop/security/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
Werkzeug==2.3.7
